requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2
feedparser
pythonping
pynewtonmath
git+https://github.com/starry69/python-markdown2.git
spongemock
zalgo-text
geopy
nltk
aiohttp>=2.2.5
Pillow>=4.2.0
PyLyrics
hurry.filesize
google_trans_new
Pafy
tldextract
PyYAML
pyowm
wikipedia
bleach
markdown2
tswift
PickleDB
youtube-dl
gtts
datetime
